+++ src/components/pages/order/Main/Basket/BasketProducts.jsx
@@ -6,10 +6,16 @@ import { findObjectById } from "../../../../../utils/array";
 import { useContext } from "react";
 
 export default function BasketProducts() {
-  const { basket, isModeAdmin, handleDeleteBasketProduct, menu } =
-    useContext(OrderContext);
+  const {
+    basket,
+    isModeAdmin,
+    handleDeleteBasketProduct,
+    menu,
+    handleProductSelected,
+  } = useContext(OrderContext);
 
-  const handleOnDelete = (id) => {
+  const handleOnDelete = (event, id) => {
+    event.stopPropagation();
     handleDeleteBasketProduct(id);
   };
 
@@ -27,8 +33,13 @@ export default function BasketProducts() {
                   : IMAGE_COMING_SOON
               }
               quantity={basketProduct.quantity}
-              onDelete={() => handleOnDelete(basketProduct.id)}
+              onDelete={(event) => handleOnDelete(event, basketProduct.id)}
               isClickable={isModeAdmin}
+              onClick={
+                isModeAdmin
+                  ? () => handleProductSelected(basketProduct.id)
+                  : null
+              }
             />
           </div>
         );
@@ -36,6 +47,7 @@ export default function BasketProducts() {
     </BasketProductsStyled>
   );
 }
+
 const BasketProductsStyled = styled.div`
   /* border: 1px solid red; */
   flex: 1;
