+++ src/utils/maths.js
@@ -1,5 +1,3 @@
-import { findObjectById } from "./array";
-
 export function formatPrice(priceToFormat) {
   let price = priceToFormat;
 
@@ -18,11 +16,3 @@ export function replaceFrenchCommaWithDot(price) {
   if (typeof price === "string") price = parseFloat(price.replace(",", "."));
   return price;
 }
-
-export const calculateSumToPay = (basket, menu) => {
-  return basket.reduce((total, basketProduct) => {
-    const menuProduct = findObjectById(basketProduct.id, menu);
-    total += menuProduct.price * basketProduct.quantity;
-    return total;
-  }, 0);
-};
